% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_postQC_postmed_rel.R
\name{calculate_postQC_postmed_rel}
\alias{calculate_postQC_postmed_rel}
\title{Function to remove samples by name}
\usage{
calculate_postQC_postmed_rel(input)
}
\arguments{
\item{input}{location of the root Symportal output}

\item{genus}{drop columns summing to zero}
}
\description{
Function to remove samples by name
}
