% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_qc_postmed_bygenus.R
\name{calculate_qc_postmed_bygenus}
\alias{calculate_qc_postmed_bygenus}
\title{Function to remove samples by name}
\usage{
calculate_qc_postmed_bygenus(input, genus)
}
\arguments{
\item{input}{location of the root Symportal output}

\item{genus}{drop columns summing to zero}
}
\description{
Function to remove samples by name
}
