% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable_seqs.R
\name{datatable_seqs}
\alias{datatable_seqs}
\title{Wrapper around DT for printing sequence tables}
\usage{
datatable_seqs(input)
}
\arguments{
\item{input}{filename for table}
}
\value{
datatable
}
\description{
Wrapper around DT for printing sequence tables
}
