% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_fasta.R
\name{extract_fasta}
\alias{extract_fasta}
\title{Function to extract fasta files from Symportal}
\usage{
extract_fasta(
  folder,
  type = "postmed",
  keep_samples = NULL,
  drop_samples = NULL
)
}
\arguments{
\item{folder}{location of the root Symportal output}

\item{type}{one of "premed" or "postmed"}
}
\value{
its_profile A data.frame of ITS profiles for each sample.ID
}
\description{
Function to extract fasta files from Symportal
}
