% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_its2.R
\name{extract_its2}
\alias{extract_its2}
\title{Function to extract its2 profile and UIDs from Symportal}
\usage{
extract_its2(folder)
}
\arguments{
\item{folder}{location of the root Symportal output}
}
\value{
its_profile_UID A data.frame of ITS profiles for each UID
}
\description{
Function to extract its2 profile and UIDs from Symportal
}
\examples{
tmp <- extract_its2(folder="/Users/rof011/Symbiodinium/20220919T102058_esampayo/")
}
