% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_its2_names.R, R/extract_its_names.R
\name{extract_its2_names}
\alias{extract_its2_names}
\title{Function to extract its2 names from Symportal}
\usage{
extract_its2_names(folder)

extract_its2_names(folder)
}
\arguments{
\item{folder}{location of the root Symportal output}
}
\value{
its_profile A data.frame of ITS profiles for each sample.ID

its_profile A data.frame of ITS profiles for each sample_name
}
\description{
.
}
