% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_its2_profile_seqs.R
\name{extract_its2_profile_seqs}
\alias{extract_its2_profile_seqs}
\title{Function to extract seqs from symportal output in long format}
\usage{
extract_its2_profile_seqs(folder, print.seqs = FALSE)
}
\arguments{
\item{folder}{location of the root Symportal output}

\item{print.seqs}{prints unique seqs output}
}
\value{
A data.frame of seq.ID (columns) and sample.ID (rows) with either relative or absolute abundance of sequences.
}
\description{
Function to extract seqs from symportal output in long format
}
