% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_its2_profile_split.R
\name{extract_its2_profile_split}
\alias{extract_its2_profile_split}
\title{Function to extract seqs from symportal output}
\usage{
extract_its2_profile_split(folder, print.seqs = FALSE)
}
\arguments{
\item{folder}{location of the root Symportal output}

\item{print.seqs}{prints unique seqs output}
}
\value{
A data.frame of seq.ID (columns) and sample.ID (rows) with either relative or absolute abundance of sequences.
}
\description{
Function to extract seqs from symportal output
}
\examples{

tmp <- extract_its2_profile_seqs(folder = "/Users/rof011/symbiodinium/20220919T102058_esampayo", print.seqs = TRUE)

tmp <- extract_its2_profile_seqs(folder = "/Users/rof011/symbiodinium/20220919T102058_esampayo", print.seqs = TRUE)

}
