% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_its2_profile_UID.R
\name{extract_its2_profile_UID}
\alias{extract_its2_profile_UID}
\title{Function to extract its2 profile UIDs from Symportal}
\usage{
extract_its2_profile_UID(folder, type = "absolute")
}
\arguments{
\item{folder}{location of the root Symportal output}

\item{type}{relative or absolute}
}
\value{
its_profile_UID A data.frame of ITS profiles for each UID
}
\description{
Function to extract its2 profile UIDs from Symportal
}
