% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_plot_colors.R
\name{extract_plot_colors}
\alias{extract_plot_colors}
\title{Function to extract colors for plots from Symportal}
\usage{
extract_plot_colors(folder)
}
\arguments{
\item{folder}{location of the root Symportal output}
}
\value{
its_profile A data.frame of ITS profiles for each sample.ID
}
\description{
Function to extract colors for plots from Symportal
}
\examples{

tmp <- extract_plot_colors(folder="/Users/rof011/Symbiodinium/20210805T112901")



}
