% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_postmed_seqs.R
\name{extract_postmed_seqs}
\alias{extract_postmed_seqs}
\title{Function to extract postmed seqs from Symportal}
\usage{
extract_postmed_seqs(folder, type = "absolute")
}
\arguments{
\item{folder}{location of the root Symportal output}

\item{type}{relative or absolute}
}
\value{
its_profile A data.frame of ITS profiles for each sample.ID
}
\description{
Function to extract postmed seqs from Symportal
}
