% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_seqs.R
\name{extract_seqs}
\alias{extract_seqs}
\title{Function to extract sequences from symportal output}
\usage{
extract_seqs(
  folder,
  metadata = NULL,
  type = "absolute",
  factors = NULL,
  onlyprofile = FALSE,
  names = FALSE,
  arrange = TRUE,
  seed = NULL,
  janitor = FALSE,
  silent = TRUE,
  ...
)
}
\arguments{
\item{folder}{location of the root Symportal output}

\item{metadata}{location of a metadata file in csv, must contain a column called sample_name with matches}

\item{type}{returns either "relative" or "absolute"}

\item{factors}{option to include a vector list of column names, e.g: c("location", "latitude") to not import every column in a metadata file, if left NULL then will import all columns}

\item{onlyprofile}{TRUE/FALSE return only profiles}

\item{names}{print names list, FALSE by default}

\item{arrange}{arrange by seq_id alphanumerically, TRUE by default}

\item{janitor}{FALSE by default, set to TRUE to use janitor clean_names on inputs}

\item{silent}{print summary statistics, FALSE by default}
}
\value{
A data.frame of seq_id (columns) and sample_name (rows) with either relative or absolute abundance of sequences.
}
\description{
extract sequences from symportal output
}
