% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_seqs_long.R
\name{extract_seqs_long}
\alias{extract_seqs_long}
\title{Function to extract sequences from symportal output}
\usage{
extract_seqs_long(
  folder,
  type = "relative",
  clade = LETTERS[1:10],
  cutoff = 1000,
  silent = TRUE
)
}
\arguments{
\item{folder}{location of the root Symportal output}

\item{type}{returns either "relative" or "absolute"}

\item{clade}{filter by single "C" or multiple clades c("C", "D") to filter sequences by}

\item{cutoff}{Set threshold to remove samples if less than the cutoff (defaults to 1000)}

\item{silent}{defaults to TRUE, if FALSE then prints a list of removed sample names}
}
\value{
A data.frame of seq.ID (columns) and sample.ID (rows) with either relative or absolute abundance of sequences. This is the same function as extract_seqs, but returns in long format not wide.
}
\description{
Function to extract sequences from symportal output
}
\examples{
## not run:
# tmp <- extract_seqs(folder="20220919T102058_esampayo", type="absolute", clade=c("C", "D"))
# tmp2 <- extract_seqs(folder="20220919T102058_esampayo", type="relative", clade="C", cutoff=100)
}
