% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_seqs_wide.R
\name{extract_seqs_wide}
\alias{extract_seqs_wide}
\title{Function to extract sequences from symportal output}
\usage{
extract_seqs_wide(
  folder,
  type = "relative",
  clade = LETTERS[1:10],
  threshold = 0,
  remove_zero = TRUE,
  drop_samples = NULL,
  keep_samples = NULL,
  drop_seqs = NULL,
  silent = TRUE,
  ...
)
}
\arguments{
\item{folder}{location of the root Symportal output}

\item{type}{returns either "relative" or "absolute"}

\item{clade}{filter by single "C" or multiple clades c("C", "D") to filter sequences by}

\item{threshold}{Set threshold to remove samples if less than the threshold (defaults to 1000)}

\item{drop_samples}{drop samples by named vector, e.g. c("H00B07", "H00B06"), or by one or more partial matches, e.g. c("07","B06")}

\item{keep_samples}{drop samples by named vector, e.g. c("H00B07", "H00B06"), or by one or more partial matches, e.g. c("07","B06")}

\item{drop_seqs}{drop seqs by named vector, e.g. c("X2777817_G", "X2777816_G"), or by one or more partial matches, e.g. c("X2","OT")}

\item{silent}{defaults to TRUE, if FALSE then prints a list of removed sample names}

\item{...}{onwards}
}
\value{
A data.frame of seq.ID (columns) and sample.ID (rows) with either relative or absolute abundance of sequences.
}
\description{
Function to extract sequences from symportal output
}
