% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_seqs.R
\name{filter_seqs}
\alias{filter_seqs}
\title{Function to extract sequences from symportal output in long format}
\usage{
filter_seqs(
  input,
  folder = NULL,
  type = "relative",
  drop_samples = NULL,
  keep_samples = NULL,
  drop_seqs = NULL,
  keep_seqs = NULL,
  keep_profiles = FALSE,
  drop_profiles = FALSE,
  random_sample_n = NULL,
  silent = TRUE,
  clade = NULL,
  threshold = 0,
  ...
)
}
\arguments{
\item{input}{Input data.frame with seq_id and sample_name}

\item{folder}{Optional, folder containing profile information}

\item{type}{"relative" or "absolute" to specify type of abundance to return}

\item{drop_samples}{Vector of samples to exclude by name or pattern}

\item{keep_samples}{Vector of samples to include by name or pattern}

\item{drop_seqs}{Vector of sequences to exclude by name or pattern}

\item{keep_seqs}{Vector of sequences to include by name or pattern}

\item{keep_profiles}{Logical, whether to keep only sequences in matching ITS2 profile}

\item{drop_profiles}{Logical, whether to exclude sequences in matching ITS2 profile}

\item{random_sample_n}{Integer, number of random samples to select}

\item{silent}{Print messages}

\item{clade}{Vector of clades to filter sequences by (e.g., clade=C, clade=("C", "D"))}

\item{threshold}{Numeric, threshold to remove samples with lower total abundance}

\item{seed}{Optional, seed value for random sampling}
}
\value{
A data.frame of seq_id (columns) and sample.ID (rows) with abundance
}
\description{
Function to extract sequences from symportal output in long format
}
