% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histo_seq.R
\name{histo_seq}
\alias{histo_seq}
\title{Test function to rapidly plot seqs}
\usage{
histo_seq(folder, random = TRUE, n, seed)
}
\arguments{
\item{folder}{location of the root Symportal output}

\item{n}{subsample sample_names by n}

\item{seed}{value for set.seed}
}
\description{
histo_seq
}
