% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_fasta_all.R
\name{import_fasta_all}
\alias{import_fasta_all}
\title{Import fasta}
\usage{
import_fasta_all(folder, type = "postmed")
}
\arguments{
\item{folder}{location of the root Symportal output}

\item{type}{returns either "premed" or "postmed"}
}
\description{
Import fasta
}
