% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_postmed_abs.R
\name{import_postmed_abs}
\alias{import_postmed_abs}
\title{Function to extract postmed sequences from symportal output}
\usage{
import_postmed_abs(folder, format = "long", ...)
}
\arguments{
\item{folder}{location of the root Symportal output}

\item{type}{"relative" or "absolute"}
}
\description{
Function to extract postmed sequences from symportal output
}
