% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_seqs2.R
\name{plot_seqs2}
\alias{plot_seqs2}
\title{Function to plot seq.IDs per sample and hierarchical clustering of sample order with dendrogram}
\usage{
plot_seqs2(input, folder, format = "ggplot", facet = TRUE)
}
\arguments{
\item{folder}{location of the root Symportal output}

\item{type}{returns either "relative" or "absolute"}

\item{clade}{filter by single "C" or multiple clades c("C", "D") to filter sequences by}

\item{drop_samples}{drop samples by named vector, e.g. c("H00B07", "H00B06"), or by one or more partial matches, e.g. c("07","B06")}

\item{keep_samples}{drop samples by named vector, e.g. c("H00B07", "H00B06"), or by one or more partial matches, e.g. c("07","B06")}

\item{drop_seqs}{drop seqs by named vector, e.g. c("X2777817_G", "X2777816_G"), or by one or more partial matches, e.g. c("X2","OT")}

\item{threshold}{Set threshold to remove samples if less than the threshold (defaults to 1000)}

\item{silent}{defaults to TRUE, if FALSE then prints a list of removed sample names}

\item{...}{pass}
}
\value{
ggplot and dendrogram for seqs/samples
}
\description{
Function to plot seq.IDs per sample and hierarchical clustering of sample order with dendrogram
}
