% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_seqs_dendro.R
\name{plot_seqs_dendro}
\alias{plot_seqs_dendro}
\title{Function to plot seq.IDs per sample and hierarchical clustering of sample order with dendrogram}
\usage{
plot_seqs_dendro(folder, type = "relative", dist = "bray")
}
\arguments{
\item{folder}{location of the root Symportal output}

\item{type}{one of "relative" or "absolute"}

\item{dist}{a distance matrix method from vegan, see ?vegdist for details}
}
\value{
ggplot and dendrogram for seqs/samples
}
\description{
Function to plot seq.IDs per sample and hierarchical clustering of sample order with dendrogram
}
