% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_samples_by_name.R
\name{remove_samples_by_name}
\alias{remove_samples_by_name}
\title{Function to remove samples by name}
\usage{
remove_samples_by_name(input, drop, remove_zero = TRUE, ...)
}
\arguments{
\item{input}{location of the root Symportal output}

\item{drop}{drop list}

\item{remove_zero}{drop columns summing to zero}
}
\description{
Function to remove samples by name
}
