% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_samples_by_seq_abund.R
\name{remove_samples_by_seq_abund}
\alias{remove_samples_by_seq_abund}
\title{Function to plot sequences from symportal output}
\usage{
remove_samples_by_seq_abund(input, value, ...)
}
\arguments{
\item{folder}{location of the root Symportal output}

\item{type}{returns either "relative" or "absolute"}
}
\description{
Function to plot sequences from symportal output
}
