% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_its2_names.R
\name{extract_its2_names}
\alias{extract_its2_names}
\title{Function to extract its2 names from Symportal}
\usage{
extract_its2_names(folder)
}
\arguments{
\item{folder}{location of the root Symportal output}
}
\value{
its_profile A data.frame of ITS profiles for each sample.ID
}
\description{
Function to extract its2 names from Symportal
}
\examples{

tmp <- extract_its2_names(  folder="/Users/rof011/Symbiodinium/20210805T112901")



}
