% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_its2_profile.R
\name{extract_its2_profile}
\alias{extract_its2_profile}
\title{Function to extract its2 profiles from Symportal}
\usage{
extract_its2_profile(folder)
}
\arguments{
\item{folder}{location of the root Symportal output}
}
\value{
its_profile A data.frame of ITS profiles for each sample.ID
}
\description{
Function to extract its2 profiles from Symportal
}
\examples{

tmp <- extract_its2_profile(folder="/Users/rof011/Symbiodinium/20220919T102058_esampayo/")
}
