% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_metadata.R
\name{extract_metadata}
\alias{extract_metadata}
\title{Function to extract metadata from symportal output}
\usage{
extract_metadata(
  folder,
  type = "relative",
  clade = LETTERS[1:10],
  cutoff = 1000,
  silent = TRUE
)
}
\arguments{
\item{folder}{location of the root Symportal output}
}
\value{
A data.frame of seq.ID (columns) and sample.ID (rows) with either relative or absolute abundance of sequences.
}
\description{
Function to extract metadata from symportal output
}
\examples{

tmp <- extract_metadata(folder="/Users/rof011/symbiodinium/20220919T102058_esampayo")
}
