% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_plot_colors.R
\name{extract_plot_colors}
\alias{extract_plot_colors}
\title{Function to extract colors for plots from Symportal}
\usage{
extract_plot_colors(folder)
}
\arguments{
\item{folder}{location of the root Symportal output}
}
\value{
list of colors for plotting
}
\description{
internal function
}
