% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_profiles_abs.R
\name{import_profiles_abs}
\alias{import_profiles_abs}
\title{Function to extract sequences from symportal output}
\usage{
import_profiles_abs(folder, metadata, factors)
}
\arguments{
\item{folder}{location of the root Symportal output}

\item{metadata}{name of metadata file (e.g. "metadata_20230120T102936.csv") within the Symportal output folder}

\item{factors}{list of columns for the metadata}
}
\description{
same as extract_seqs but in long format
}
