% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_profiles_rel.R
\name{import_profiles_rel}
\alias{import_profiles_rel}
\title{Function to extract sequences from symportal output}
\usage{
import_profiles_rel(folder)
}
\arguments{
\item{folder}{location of the root Symportal output}

\item{type}{returns either "relative" or "absolute"}
}
\description{
same as extract_seqs but in long format
}
