% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_postmed_rel.R
\name{plot_postmed_rel}
\alias{plot_postmed_rel}
\title{Function to plot sequences from symportal output}
\usage{
plot_postmed_rel(
  folder,
  plotly = FALSE,
  format,
  metadata,
  factors,
  facet = NULL,
  ...
)
}
\arguments{
\item{folder}{location of the root Symportal output}

\item{format}{returns either "relative" or "absolute"}

\item{metadata}{name of metadata file (e.g. "metadata_20230120T102936.csv") within the Symportal output folder}

\item{factors}{list of columns for the metadata}

\item{...}{passes functions}

\item{order}{order by factor level}
}
\description{
Function to plot sequences from symportal output
}
