% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_postmed_rel.R
\name{plot_postmed_rel}
\alias{plot_postmed_rel}
\title{Function to plot sequences from symportal output}
\usage{
plot_postmed_rel(folder)
}
\arguments{
\item{folder}{location of the root Symportal output}

\item{type}{returns either "relative" or "absolute"}
}
\description{
Function to plot sequences from symportal output
}
